requirements.txt
langchain_openai
langchain_core
python_dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
cassio
beautifulsoup4
langchain-groq
sentence_transformers
ctransformers
python-box
unstructured
tiktoken
pinecone-client
openai
pandas
numpy
langchain_pinecone